Explicit type conversions

+++ src/Monitor/Monitor.php
@@ -28,7 +28,7 @@ class Monitor
      *
      * @return \stdClass
      */
-    public function info() : stdClass
+    public function info(): stdClass
     {
         // preparing cpu info
         $uname = shell_exec('uname -r ');
@@ -58,11 +58,11 @@ class Monitor
      *
      * @return \stdClass
      */
-    public function hardware() : stdClass
+    public function hardware(): stdClass
     {
-        $output = shell_exec('cat /sys/class/thermal/thermal_zone0/temp');
+        $output = (float)shell_exec('cat /sys/class/thermal/thermal_zone0/temp');
         $temp = round(($output) / 1000, 1);
-        $output = shell_exec('echo "$(</proc/uptime awk \'{print $1}\')"');
+        $output = (float)shell_exec('echo "$(</proc/uptime awk \'{print $1}\')"');
         $timeAlive = StringHelpers::secondsToTime((int)$output);
 
         // data object
@@ -81,7 +81,7 @@ class Monitor
      *
      * @return \stdClass
      */
-    public function loadAverage() : stdClass
+    public function loadAverage(): stdClass
     {
         $output = shell_exec('uptime');
         $loadavg = explode(' ', substr($output, strpos($output, 'load average:') + 14));
@@ -109,7 +109,7 @@ class Monitor
      *
      * @return stdClass
      */
-    public function memory() : stdClass
+    public function memory(): stdClass
     {
         //$mem_free = (int)shell_exec("free -m | awk '/buffers\/cache/ {print $3}'");
         $memTotal = (int)shell_exec("free -m | awk '/Mem/ {print $2}'");
@@ -169,7 +169,7 @@ class Monitor
      *
      * @return stdClass
      */
-    public function network() : stdClass
+    public function network(): stdClass
     {
         $output = shell_exec('sh ' . __DIR__ . '/transfer_rate.sh');
         $rates = explode(' ', $output);
@@ -189,7 +189,7 @@ class Monitor
      *
      * @return \stdClass
      */
-    public function storage() : stdClass
+    public function storage(): stdClass
     {
         $output = shell_exec('df -H');
         $table_rows = preg_split('/$\R?^/m', $output);
@@ -207,7 +207,7 @@ class Monitor
      *
      * @return array
      */
-    private function prepareColumns(string $row) : array
+    private function prepareColumns(string $row): array
     {
         return array_values(array_filter(explode(' ', $row), 'strlen'));
     }
